factor: Move recursive factoring logic from rho to factor
No functional change, but prepares a coming optimisation.
